Updated openstack/openstack Project: openstack/keystone  6495028090ca8da81aabd7d0768bb9cfcebe8c5c
Make controllers call the new, split out, role manager

This is the part of the more comprehensive split of
assignments, which rationalizes both the backend and controllers.
In order to make this change easier for reviewers, it is divided
into a number of smaller patches. Previous patches:

- Moved role management into its own manager and drivers
- Fixed incorrect doc strings for grant driver methods

This patch updates any controllers that make calls to manage
roles to use the new role manager.

Future patches will:

- Update the tests to call the new role manager
- Refactor assignment manager and driver methods to logically
  separate project/domains from the actual assignments
- Split projects and domains into their own backend
- Split the controllers so they call the correct manager
- Update the tests to call the new correct manager

Partially implements: bp pluggable-assignments

Change-Id: Ia8777cd44921372afc8e918e2ae58d760f0c96fa