Updated openstack/openstack Project: openstack/nova  a9fca6291401ba0095299c2748f7f4a0c1026ac2
Fix OverQuota headroom KeyError in nova-network allocate_fixed_ip

Commit 0045f7d50cad345c26ae02394ad57716b270d268 moved the headroom
calculation out of the DB API and into the compute API for when
instances go over quota (more specifically flavor information for an
instance create), but didn't update the nova-network manager code to
remove the usage of the headroom kwarg from the OverQuota exception
for logging.

This removes the usage of 'headroom' and also fixes the logging a bit so
the 'used' value makes more sense, where used = in_use + reserved.

Closes-Bug: #1410310

Change-Id: I24802e8b5dfe7741b505696c885871a04f769b12