Updated openstack/openstack Project: openstack/taskflow  1ed0f22fd35c4e08667f21bf549b3fe84a230d32
Use constants for runner state machine event names

Instead of using strings it is better if we can use constants (that
may be the same/adjusted strings) and use those instead in the state
machine used in the runner.

The names are adjusted (and the state graph diagram and docstring)
to reflect names that fit better with there intended meaning and
usage.

Change-Id: Iaf229d6e37730545ba9f2708d118697cb7145992