Updated openstack/openstack Project: openstack-dev/devstack  116f9f826effd3682c8b70cde8550727161c3384
Allow specification of ironic callback timeout

We need to be able to adjust Ironic's timeout for a node's deployment
callback.  The default is much longer than Tempest's server build
timeout, causing Ironic deployment issues to be masked by more generic
server timeouts in the tempest logs.  Being able to set this to be
lower than the nova timeout, we'll have Ironic errors that we can
fingerprint.

Change-Id: I7b8eeda504da7ffd64967bbcfa2625acf418f263
Related-bug: #1408067