Updated openstack/openstack Project: openstack/nova  f9ad04d3808a2a7dbd7b26db6823276f69b6f4e5
libvirt: set guest cpu_shares value as a multiple of guest vCPUs

Currently, a default value of cpu_shares is being assigned to all
guest instances, regardless of the number of vCPUs has been
configured for each guest. As cpu_shares are specifying the weight
to provide CPU time, this reduces the amount of time each
vCPU gets, for instances with a higher number of vCPUs.

Setting the default value of guest cpu.shares to be a multiple of
the number of guest vcpus.

Closes-Bug: #1383377

Change-Id: I3a669da3536061919e3673da88c102a868999902