Updated openstack/openstack Project: openstack/cinder  42c9f249a79adf18a044d49502600877e58bd3ef
Move oslo.db to oslo_db namespace

This is the second in a series of changes to move to using
the new oslo_<library> namespace that is being used for
oslo libraries.

There is currently a shim in place that is allowing the old
oslo.<library> imports to work, but we need to be prepared for
when the shims go away.  Thus, we need patches like this one to
move to the new namespace.

This patch also updates our hacking check to ensure that no instances
of oslo.db sneak back in.

Change-Id: Ibebed114a362e199e03c83d7d229841e8f88932b