Updated openstack/openstack Project: openstack/neutron  f3505e873f59702fad1196cafdd224eea72ba674
bigswitch: drop usage of SecurityGroupServerRpcApiMixin

Drop usage of SecurityGroupServerRpcApiMixin in the bigswitch plugin.
This is required to be able to eventually move this API into a
messaging namespace.  It needs to use its own messaging client
instance, instead of a different one it gets after being used as a
mixin.

Part of blueprint rpc-docs-and-namespaces.

Change-Id: Ib1c06d5b79c51fb63131efdf6ed83a7e85a7b820