Updated openstack/openstack Project: openstack/neutron  21b20b7e2d4a4ff1a18ef2b19508cab7cca9f27e
hyperv: drop usage of SecurityGroupServerRpcApiMixin

Drop usage of SecurityGroupServerRpcApiMixin in the hyperv plugin.
This is required to be able to eventually move this API into a
messaging namespace.  It needs to use its own messaging client
instance, instead of a different one it gets after being used as a
mixin.

Part of blueprint rpc-docs-and-namespaces.

Change-Id: Idad9732be1c225262b39e07455d41bd61477c39d