Updated openstack/openstack Project: openstack/nova  e2609e961df1ea25eeddafad39b3c3a15d9f3a6b
xenapi: don't send terminating chunk on errors

If we send the terminator on an error, it looks like we have actually
completed the upload, when that was not actually the case, we have
failed, and ideally we want glance to notice and mark the snapshot as
failed.

This partly reverts the following commit:
e039b036b5e9dbaff8b37f7ab22c209b71bdc182

It is not a full revert of the above commit, and some of the refactoring
and error handling changes are still useful.

Closes-Bug: #1398826
Change-Id: Ib608a32556135a5b7e666d02c4ee77722ea6e07c