Updated openstack/openstack Project: openstack/taskflow  320955672dafe531533971f4a77fcdae5a7f5a8c
Remove 'SaveOrderTask' and test state in class variables

Instead of saving task state in a class variable that is later
introspected by further test code just remove that concept (which
doesn't work in multiprocessing or worker engines which can not
have access those types of shared/globally available concepts due
to how they run) and use a specialized listener that can gather the same
information in a more decoupled manner (and it will work in multiprocessing
and worker engines correctly).

This allows our engine test cases to work in those engine types which
increases those engines test coverage (and future coverage and engine tests
that are added).

Fixes a bunch of occurrences of bug 1357117 as well that were removed
during this cleanup and adjustment process...

Change-Id: Ic9901de2902ac28ec255bef146be5846d18f9bfb