Updated openstack/openstack Project: openstack/nova  5bbb97f7bee56fc7bb27f4de1d6c35e217466af9
Make numa_usage_from_instances consider CPU pinning

If instance exposes any CPU pinning information, it will be taken into
consideration when calculating NUMA cell usage of the host.

We also add some convenience methods to NUMACell object for pinning and
unpinning CPUs, that will raise an error if the data is inconsistent,
and a wrapper for checking weather CPU pinning was actually requested
(to avoid an error prone 'is not None' in places where the check is
needed).

Change-Id: I43543c526d3733572f4185fbc9aa2207e9b2154c
Blueprint: virt-driver-cpu-pinning