Updated openstack/openstack Project: openstack/oslo-incubator  09e6f70ee301c643b3e0103afd15a0785f531bed
Replace the 'release_notes' shell script with a python variant

Instead of using shell script (and the errors prone to its usage)
just use a nice little python script that can provide the same
template and output using jinja2, parawrap (for nicely formatted
paragraphs) and using subprocess so that all return codes are
checked.

This also uses argparse so that a nice help string is generated
and handles the README.rst file not existing or the setup.py file
called upon not existing and so-on.

Change-Id: I2e2969d2188da3c3c33ddb5adc21bdeaff3423f5