Updated openstack/openstack Project: openstack/cinder  d798294388a53c0153ff8d8a2cf3282de6eeee02
Fixed server name being retained after detach in LeftHand

There was an issue with the LeftHand driver where the
initiator names of servers were not being removed when the last
volume attached to that server was detached.

This fix will now remove the server from the LeftHand backend
when no more volumes are attached to it.

Closes-Bug: #1353137
Change-Id: I58be39d14979e1d3966232095cb49693c1424f8a