Updated openstack/openstack Project: openstack/cinder  274cf277f830a092e69636a92bc3ceb6a0f2a7e4
Add support for manage/unmanage volume commands to PureISCSIDriver

This change adds implementations of the required methods to make
manage and unmanage commands work with the PureISCSIDriver. You can now
call them and specify a volume name of an existing non-cinder volume on
the array and it will rename the volume accordingly to allow cinder to
manage it. When a volume is unmanaged it is renamed with a “-unmanaged”
suffix and left on the array.

Implements: blueprint pure-iscsi-manage-volume
Change-Id: I2fa63ba50a367fd06a2ede7768eb338bf6a1ee4e