Updated openstack/openstack Project: openstack/tempest  83feb46843cb97ab99c0eb21feb5be8e8f86c2e8
Move tempest_roles options to auth group

This commit moves the tempest_roles option to the auth group. This
option was previously in the identity section which was the incorrect
place for it. The identity group is used for config settings related
to keystone, while the auth section is used for tempest specific user
auth.

Change-Id: Id89c7c265b648cf31b3f2e684d4467a7c8c2f554