Updated openstack/openstack Project: openstack/nova  91fa3efb24b311e3cf902d0b901ed2f9316f8db9
Fixes Hyper-V boot from volume live migration

Live migration fails on Hyper-V when boot from volume is used with
CoW, as the target host tries to cache the root disk image in
pre_live_migration, but in this case the image_ref is empty.

This patch adds a check to handle the empty image_ref case.

Co-Authored-By: Claudiu Belu <cbelu@cloudbasesolutions.com>
Co-Authored-By: Adelina Tuvenie <atuvenie@cloudbasesolutions.com>

Change-Id: I60cb60ccaeb0cb8c536906d897249e31ae396923
Closes-Bug: #1362075