Updated openstack/openstack Project: openstack/nova  710ae293b98209d8e60b5e2e79c915715b3a4533
Add support for clean_shutdown to resize in compute api layer

Change Iec3dfd17725440958aac395ebc471e51afd6522e added
clean_shutdown support to the compute manager in Juno.
This change is part of a set to extend that to the compute
API layer, which will in turn allow us to extend the REST API
to provide a forced-shutdown option to those operations.

Change-Id: Ie93c3c4aa30b16e277628e1a741e46a37b794a75
Partially-Implements: blueprint user-defined-shutdown