Updated openstack/openstack Project: openstack/nova  a1a71042b4c6eadb29e6232ea463e73acf54aefa
Reduce complexity of the _get_guest_config method.

The _get_guest_config() method of libvirt's driver.py is currently the
worst offender in nova. This patch refactors the method to reduce the
complexity of the method from 40 (originally; it has since been dropped
to 35 by other patches) down to 24. With this change, the highest
complexity method is 37, so we can lower the limit in tox.ini to 38.

Change-Id: I05de301bc3dfd84ed15ce3cbb8a7c50f2d1385ed