Updated openstack/openstack Project: openstack/nova  2e9d31cf56e0085d2105c49ada3733b00f356c33
Remove lock files when remove libvirt images

When fetches an image from Glance, LibvirtDriver creates a
lockfile that has a name corresponding to the image.
It is not deleted automatically even if the image is delete
from nova. So lockfiles may accumulate in long term.
Although this may be not so harmful, a little bit ugly.

So this patch removes the external lock files when the instance
image file is deleted.

Change-Id: Ia00388cc5a9dca84df8eea038d693f37940d494e
Partial-Bug: #1256306