Updated openstack/openstack Project: openstack/heat  cfc83ee12563f9f2aaf65b27d3b427dd6d680d41
Return available info for uncreated resource

The resource show API verified that a resource is defined in stack
and has an id associated with it. Because of this, for a valid stack
resource which is not persisted, the resource show API returned the
error 'stack or resource not found'.
Fixed by validating only that the resource is defined in a stack
and return the available information in the resource show API
response.

Change-Id: Ic0a031294f4f1dffc3c9978014480d15fd2fa27b
Closes-bug: #1402531