Updated openstack/openstack Project: openstack/tempest  47278ecb11db387a82ed9793d98fcebb62e472e3
Use create_test_server in test_max_image_meta_exceed_limit

When this fails it's just a MismatchError in the logs so it's hard to
fingerprint this kind of race failure in logstash, especially since the
name of the test server is just 'test'.

Change to use the create_test_server method from the base class so the
instance name has the class name in it which would make this easier to
fingerprint in the result of an intermittent failure.

This is doubly useful since if this fails we're leaking a server
resource. By using create_test_server, if the assertion fails then at
least the base class is tracking the server created and will delete it
on tear down.

Related-Bug: #1411708

Change-Id: Ib015442a4d89dfcbb68438fe31d6fda719ddd011