Updated openstack/openstack Project: openstack/tempest  4857696ad79572c937ac132567b9fbc9b0be6a75
Add compute_quotas lock fixture to test_absolute_limits_negative

This has a negative test which tries to create a server with a number of
metadata items that exceed the quota limits but it intermittently
failing because the server gets created successfully (returns the 202).

This is most likely due to the quota class test which is making the
default quotas unlimited when it runs so we can hit a window where this
negative test will fail.

Anything compute tests in tempest which relies on quota checks for test
results should be using the compute_quotas lock fixture, so add that
here.

Closes-Bug: #1411708

Change-Id: Ic701b0dff2087002243d1ebc2ba8b4bd628d5eab