Updated openstack/openstack Project: openstack/nova  794700558efe07044d4673f2d60af00200e18f56
VMware: ensure that correct disk details are returned

When querying the root disk and a ephemeral disk is attached
we need to ensure that the correct disk details are returned.
Prior to the ephemeral disk support, the root disk was the only
disk in the directory.
We need to make sure that the disk is the root disk.

This breaks snapshots of images and resizing.

Either the root disk details will be returned, if requested, or the
last disk that is found.

Change-Id: I6d9f0c920409c5f6f53f552390fcd0d07ebf7d0b
Closes-bug: #1408613