Updated openstack/openstack Project: openstack-dev/devstack  4f154070e4688f60d88b4807822e481d36a009f7
Disable shallow cloning with GIT_DEPTH=0

The recent patch to enable shallow cloning added a method to skip
it by checking if GIT_DEPTH was null. However, this could never
be triggered because if the user specified it as null, the default
setting code would take over and set it to 1.

This patch allows the user to specify GIT_DEPTH=0 to skip the
shallow clone.

Change-Id: I00ea7ab54ed51dc3ede9ceb9ff0f11575a035d1c