Updated openstack/openstack Project: openstack/python-neutronclient  3e41128580bdf212a6680508fd9bc6104cdcce35
Add parser options for port-update and port-create

Added parser options for port parameters that can be updated
thus making the help reference for port-update more verbose.

This patch also allows subnet_id and ip_address parameters
of the --fixed-ip argument to be passed in one update request.
Added a unit test for this.

Also added a missing device_owner argument to both port-update
and port-create commands.

Change-Id: Id78a1400b282b34ee7b30437390be954a2c34f88
Closes-Bug: #1165171
Closes-Bug: #1208427
Closes-Bug: #1399681