Updated openstack/openstack Project: openstack/neutron  4483062d77f50abbdb24e5148013dec3ec04cdb6
make delete_router send delete_port to core_plugin

Have delete_router API call also delete the gateway port if needed.
Currently, the db_base_plugin_v2 method is called, bypassing the
core_plugin overload of this method. The consequence is that,
in ML2, mechanism drivers are not aware of the port deletion.

In this patch, tha gateway port is deleted through a common
method. This method checks if floating IP and VPN are attached
to the router, so those checks are removed from the delete_router
method.

Closes-Bug: #1361540

Change-Id: Iae98f3fe89126a76f16ed9c5230ce299a09ce8d8