Updated openstack/openstack Project: openstack/nova  8232b618287663fae4e83642b07a490f1bb29103
don't use debug logs for object validation

There is a curious use of debug logging to validate extensions to the
API, which is both overly verbose, and really error prone. This
instead creates a dedicated __repr__ as well as is_valid() on
extensions to separate these concerns (and remove a lot of pretty
useless debug logging in the process).

Part of bp:log-guidelines

Change-Id: I0a67499a42863661bcf4a8b959c3492575b3acb6