Updated openstack/openstack Project: openstack/taskflow  2d9b5a8d92b4e8b5499b949782f1aa1f81d03a5c
Move the persistence base to the parent directory

In order to match the directory/module layout of the
other pluggable backends better move the persistence
base module that defines the base abstract classes up
into the parent directory.

This makes it easier to look at the taskflow code-base
and understand the common layout.

Change-Id: I7887cb0241b8fe65cbdfee32c101c3df5f05d27c