Updated openstack/openstack Project: openstack/nova  3f8b2ca94f8e07ae9a5c344854f5944d39fa2efb
clean up extension loading logging

The use of AUDIT on every extension loaded (after it was loaded) is
substantial over kill. This replaces that with a single INFO message
listing the extensions that were loaded at the end of the
extension/plugin loading process. Providing equivalent info in a much
smaller package. The extensions are listed by alias (as they were in
AUDIT) and alphabetically sorted to will be much easier to see if an
extension is in or out of the list.

This also removes the use of WARN for not loading an extension because
of configuration (blacklist or whitelist) the configuration explicitly
set a thing, so this is not a warn condition. Providing the loaded
extensions list at INFO will let people determine if something is
correctly loaded or not.

It also removes the AUDIT message of entering the extension
manager. There remains a ton of DEBUG in the initial load of all this
code so this should not impact debugging.

Part of bp:log-guidelines

Change-Id: Ic43391d10630359c09f344a62ea12a22457ea9d0