Updated openstack/openstack Project: openstack/trove  ec2761a7900a87d7b2d73c6356f01b427c2979b0
Address predictable temp file vulnerability

This change addresses a predictable temporary file vulnerability in
the code path that writes the cassandra configuration. Unit tests have
been added.

Since there is a problem with Mock()'ing a shared entry point (like
os.unlink or utils.execute_with_timeout) there is an inherent
instability in these tests. The safe way around this is to make
write_config() accept arguments that can be Mock()'ed for the purpose
of testing.

A lengthy explanation of the rationale for this change is on the
openstack-trove IRC channel at about 10:40AM on 12/29/2014.

Also, there was a redundant test that has been eliminated.

Change-Id: I760b937b6714b2b2b366cd8bdb700bece6055fba
Closes-Bug: #1398195
Partial-Bug: #1398966