Updated openstack/openstack Project: openstack/horizon  a38b72813674e02bfe91c6338917f144c924b11b
Pseudo translation tool

A tool to allow pseudo translations to be created in order to identify
potential translation problems. To use the tool:
Make sure your English file is up to date:
./run_tests.sh --makemessages

Run the pseudo tool to create pseudo translations:
./run_tests.sh --pseudo de

Compile the catalog:
./run_tests.sh --compilemessages

Run your dev server. Log in and change to the language you pseudo translated.
It should look weird. More specifically, every translatable string is going
to start and end with a bracket and they are going to have some added
characters. For example, "Log In" will become "[~Log In~您好яшçあ]"
This is useful because you can inspect for the following:
- If you see a string in English it's not translatable. Should it be?
- If you see brackets next to each other that might be concatenation.
- If there is unexpected wrapping/truncation there might not be enough
  space for translations
- If you see a string in the proper translated language, it comes from an
  external source. (That's not bad, just sometimes useful to know)
- If you get new crashes, there is probably a bug. :-)

Implements blueprint: pseudo-translation-tool

Change-Id: If97754c2d4234b12b3d73616ff60527f6ad82d55