Updated openstack/openstack Project: openstack/keystone  861ce9c998952c12977d0b56fe13a32cad734895
Refactor assignment manager/driver methods

This is the part of the more comprehensive split of
assignments, which rationalizes both the backend and controllers.
In order to make this change easier for reviewers, it is divided
into a number of smaller patches.

Previous patches have:

- Moved role management into its own manager and drivers
- Fixed incorrect doc strings for grant driver methods
- Updated controllers to call the new role manager
- Updated unit tests to call the new role manager

This patch refactors the assignment manager and driver methods
to logically separate the project/domain functionality from
that of the actual assignments in preparation for the move of
projects/domains to their own backend. This patch should not
cause any functional change.

Future patches will:

- Split projects and domains into their own manager/backend
- Update the assignment controller to call the new manager
- Update the tests to call the new manager
- Split the assignment controller, giving projects/domains
  their own controller

Partially implements: bp pluggable-assignments

Change-Id: Ida92571620c1cb3f2cdf896125865c92301b1c3b