Updated openstack/openstack Project: openstack/horizon  2283b73425c91be3e1284889b30dc6c61cde7999
List port for distributed router interface

This patch fixes an issue that occurs when a distributed
virtual router (DVR) is in use on a network and a user
attempts to assign a floating IP address to an instance.

The existing approach assumes that only standard routers,
not DVR based router(s), are in use on the network. As a
result, if you attempt to assign a floating IP address,
there are no candidate ports shown in the select to bind
the floating IP address to. The dialog that allows you
to attach a floating IP address will show a message in
the "Port to be associated" select of "No ports available".

This patch ensures that if a distributed router exists
for the network, the appropriate port for that router
is displayed in the list of ports.

Change-Id: I2a23389aca02ea8ee2cc8970e2a264d2e8c41dae
Closes-Bug: 1388305