Updated openstack/openstack Project: openstack/nova  31d0f5acc6ff1343e323f932274b27305e6acb02
Provide compatibliity for db.compute_node_statistics

As some computes are providing the host field and others are not, let's modify
how we calculate the statistics by merging both versions.

Note: This change won't impact how the API is getting stastistics as the result
is not a ComputeNode or a list of ComputeNodes but instead a mix of stats.

Partially-Implements blueprint detach-service-from-computenode

Change-Id: I889efc41babb2c82018122b5f255eb0fca9b452b