Updated openstack/openstack Project: openstack/heat  12a717f926ce17091db558c66aad007e5a6fd7aa
Deprecate QUEUE_ID in heat_zaqar

ZaqarQueue in heat_zaqar has 2 ways to receive the ID of the queue:
- use get_attr(QUEUE_ID)
- use get_resource <resource_name> or Ref
The first way has been deprecated because in common practice
it had better use get_recources for receive an ID of the resource.
In addition, this attribute has been excluded from auto-tests
for ZaqarQueue.

Change-Id: I43d07401b580555fe5bfa443836e60a708ce54bf