Updated openstack/openstack Project: openstack-dev/devstack  2d7c346f0c8cefe622edc8fed47cc2782aeb9a9b
Exit when printing error about database config

When finding no configured database, devstack prints
a suitable error in the logs

   "No database enabled"

but then just carries on running, as a result the developer
is very unlikely to ever see the original error message.
Change the 'echo' to a 'die' so that it stop immediately
making the error message clearly visible.

Change-Id: Ibd86bfcb5d4a3b90a1ee7a5bd637b01124e3a6ba