Updated openstack/openstack Project: openstack/neutron  ffb55a9f5fdce01c240e3953a58fe4dd45eef04c
nec: drop usage of SecurityGroupServerRpcApiMixin

Drop usage of SecurityGroupServerRpcApiMixin in the nec plugin.
This is required to be able to eventually move this API into a
messaging namespace.  It needs to use its own messaging client
instance, instead of a different one it gets after being used as a
mixin.

Part of blueprint rpc-docs-and-namespaces.

Change-Id: Id76c3fd22a0e847660f58e4d3c8a1001e3654918