Updated openstack/openstack Project: openstack/neutron  c1b4a22230d3bbd08a46552aa4eda4ddb48c9376
ofagent: drop usage of SecurityGroupServerRpcApiMixin

Drop usage of SecurityGroupServerRpcApiMixin in the ofagent plugin.
This is required to be able to eventually move this API into a
messaging namespace.  It needs to use its own messaging client
instance, instead of a different one it gets after being used as a
mixin.

Part of blueprint rpc-docs-and-namespaces.

Change-Id: I932072bd5ab40c58c0706264cf5087d4f5786fce