Updated openstack/openstack Project: openstack/neutron  d640189d3536f66cf94f0caf8139877ca73c7616
oneconvergence: drop usage of SecurityGroupServerRpcApiMixin

Drop usage of SecurityGroupServerRpcApiMixin in the oneconvergence
plugin.  This is required to be able to eventually move this API into
a messaging namespace.  It needs to use its own messaging client
instance, instead of a different one it gets after being used as a
mixin.

Part of blueprint rpc-docs-and-namespaces.

Change-Id: I635fbeb57f4d5ee8a7ce4ca353a32b543e7d2433