Updated openstack/openstack Project: openstack/neutron  af22bbc729578982551e5d99c5f21fe3b5fa9380
ovs: drop usage of SecurityGroupServerRpcApiMixin

Drop usage of SecurityGroupServerRpcApiMixin in the ovs plugin.
This is required to be able to eventually move this API into a
messaging namespace.  It needs to use its own messaging client
instance, instead of a different one it gets after being used as a
mixin.

Part of blueprint rpc-docs-and-namespaces.

Change-Id: I6c3f3325b4337e55aed2316dade8025425c4e919