Updated openstack/openstack Project: openstack/nova  31c04da820dac9603fbdec840f4301aff9a1d5be
Get settable user quota maximum correctly

Get_user_quota() has a side effect to the argument 'user_quota'. When
getting a settable quota range for a user, this method is called by
get_settable_quotas(). Get_settable_quota() does not expect that the
project quota is set when the user quota is not set. The side effect
makes the settable maximum value incorrectly great.

In this patch, to avoid this side effect and an extra DB access,
get_user_quota() copies the 'user_quota'.

To the unit test, the patch adds the case when the project quota for
floating_ip resource is set. The test mocks _process_quotas() instead
of get_user_quotas() to check the side effect of get_user_quotas().

Change-Id: I13f875740d6c5ff82dbd0203d42987e10b38b1df
Closes-Bug: 1400945