Updated openstack/openstack Project: openstack/swift  9e7c536ca938cf1fa87ed5549d60f057076244ef
Implement OPTIONS verb for storage nodes.

Many times new deployers get mysterious errors after first setting up their
Swift clusters. Most of the time, the errors are because the values in the ring
are incorrect (e.g. a bad port number). OPTIONS will be used in a ring checker
(which is WIP) that validates values in the ring.

This patch includes OPTIONS for storage nodes and respective tests.

Change-Id: Ia0033756d070bef11d921180e8d32a1ab2b88acf