Updated openstack/openstack Project: openstack/requirements  3e98b47870d8a9deb50d3e9e26c2ac0de7036cb8
Bump gabbi to 0.5.0

More in depth use with ceilometer and gnocchi revealed that
establishment of the wsgi-intercept must be delayed until
after fixtures have had a chance to run. The fixtures do things
like establishing oslo_config.cfg which some WSGI app factories
need immediately.

Fixtures are also capable now of causing an entire TestSuite to be
skipped if some prerequiste cannot be satisfied (for instance a
database in not available). This has made it possible to easily
integrate the gabbi-drive tests on the same tox targets as other
unit tests.

Change-Id: Ie4e2b6a1c06e2148951a9029ce473e2c820b544e