Updated openstack/openstack Project: openstack/neutron  63120b23fd2f0431cde8bd13a2f28e5d5e0f7478
SIGHUP keepalived if L3 agent restarts

Currently restarting the L3 agent does not SIGHUP any running
keepalived processes. This is an issue if the L3 agent crashes
and is then restarted by an external tool. In this case, the L3
agent resyncs with the controller but does not SIGHUP keepalived.

It's also an issue during upgrades.

Closes-Bug: #1412542
Change-Id: Iec29b483f499f02c593f6e0663c7f3504c615c1a