Updated openstack/openstack Project: openstack/nova  0aa82557eaaac1ef8a1a6f49bdab4cf97b152ded
Switch to oslo.vmware API for reading and writing files

This started off as an exercise to remove httplib.HTTPSConnection
to prevent MiM attacks but morphed into a full fledged replacement
of code in read_write_util to use functionality from oslo.vmware.

As a side-effect, we remove the use of the HTTPSConnection and
are switching over to requests library as well.

Closes-Bug: #1374000
Change-Id: I917c34042c501af03725b0504542e00e7d80e511