Updated openstack/openstack Project: openstack/trove  7270daff4a0246279a4116111c60e728a159bec8
Correct calls to mask_password() which no longer work

There were some places in the code where the mask_password() routine
was being called incorrectly and these now resulted in a failure due
to a newer version of oslo.utils (strutils).

I've fixed the one location which was causing the reported failure and
seven other places where the same incorrect code was present.

Change-Id: I69f47806f1552358071b031a9ef68f016df0e7c2
Closes-Bug: 1413677