Updated openstack/openstack Project: openstack/python-cinderclient  f26de0f8fc698604bf272645ad53daa41fc2ef63
cinder list fails with 'name' sort key

The client restricts the sort keys that the user can supply. The
'name' key is allowed but is not the correct key for sorting by name,
it should be 'display_name'.

If 'name' is used then the client returns with the error 500
Internal Server Error.

This patch will add support for mapping client sort keys (eg, 'name')
to server sort keys (eg, 'display_name'), allowing the user to supply
the 'name' key to sort by name.

This patch also adds UT for the sort key and direction error cases
when an invalid value is supplied.

Change-Id: I0bdad6d61da83a3924a6b18678afe4722b5778d6
Closes-Bug: 1404020