Updated openstack/openstack Project: openstack/neutron  08438a6d3660f263fdf65b45c8c177a54f088ae7
linuxbridge: untangle SecurityGroupAgentRpcMixin

This patch separates the use of SecurityGroupAgentRpcMixin out to its
own class.  This matches most of the rest of the code base.  This
separation is needed to be able to eventually move this rpc API into
its own messaging namespace.  Now that it's separate, a future change
can pass the new class an instance of SecurityGroupServerRpcApi
instead of assuming that the LinuxBridgePluginApi includes
SecurityGroupServerRpcApiMixin.

Part of blueprint rpc-docs-and-namespaces.

Change-Id: I913ef0e2f981b258ea03a519f53a55b4afe80daa