Updated openstack/openstack Project: openstack/horizon  c97b81f5b5454487956ff9ae7fbfd0e178ff21da
Updating startpanel for pluggables

'run_tests.sh -m startpanel <name>' command for creating new panels
still requires specifying the dashboard with the
'--dashboard=<dash-python-path>' otherwise it fails.

With the new mechanism for plugging in extensions, this should no longer
be a requirement as the handling of the enabled file will do the
registration.

It is still a valid option for integrated content generation though.

Closes-Bug: #1410025
Change-Id: Ibaaa6c4d36defeae420919d0ea647e18127afefb