Updated openstack/openstack Project: openstack/neutron  682d35e5fad7c2515f6138729dc9c86c649aef4e
linuxbridge: drop usage of SecurityGroupServerRpcApiMixin

Drop usage of SecurityGroupServerRpcApiMixin in the linuxbridge agent.
This is required to be able to eventually move this API into
a messaging namespace.  It needs to use its own messaging client
instance, instead of a different one it gets after being used as
a mixin.

Part of blueprint rpc-docs-and-namespaces.

Change-Id: I736468ebd32822352e44eda8c5f1a25972a70fda