Updated openstack/openstack Project: openstack/neutron  8d5c881dec511038b3540a94f9c7c5d0667abbba
mlnx: untangle SecurityGroupAgentRpcMixin

This patch separates the use of SecurityGroupAgentRpcMixin out to its
own class.  This matches most of the rest of the code base.  This
separation is needed to be able to eventually move this rpc API into
its own messaging namespace.  Now that it's separate, a future change
can pass the new class an instance of SecurityGroupServerRpcApi
instead of assuming that the PluginApi instance includes
SecurityGroupServerRpcApiMixin.

Part of blueprint rpc-docs-and-namespaces.

Change-Id: Ie850a262bc7e4b697003cf6d91780dc819c688e0