Updated openstack/openstack Project: openstack/neutron  d40fa6ceb4cbe8555fcdb96dd79dedbd1817e52e
mlnx: drop usage of SecurityGroupServerRpcApiMixin

Drop usage of SecurityGroupServerRpcApiMixin in the mlnx agent.
This is required to be able to eventually move this API into
a messaging namespace.  It needs to use its own messaging client
instance, instead of a different one it gets after being used as
a mixin.

Part of blueprint rpc-docs-and-namespaces.

Change-Id: Iba96bddf30f33569a96548a5f116d66a4db6fc8f