Updated openstack/openstack Project: openstack/neutron  c991a19f6f4a58769d0e1da8fd4697d593003dbb
sriovnicagent: drop usage of SecurityGroupServerRpcApiMixin

Drop usage of SecurityGroupServerRpcApiMixin in sriovnicagent.
This is required to be able to eventually move this API into
a messaging namespace.  It needs to use its own messaging client
instance, instead of a different one it gets after being used as
a mixin.

Part of blueprint rpc-docs-and-namespaces.

Change-Id: I157866bb65f884f80a9c3d70ffdcd487644caaef